pyproject.toml
[project]
name = "multi-agents-cykg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-neo4j>=0.5.0",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "mcp-use>=1.3.7",
    "numpy>=2.3.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rdflib>=7.1.4",
]
